requirements.txt
# Market Data
yfinance==0.2.36
nsepy==0.8

# Database
SQLAlchemy==2.0.25
alembic==1.13.1

# Scheduling
APScheduler==3.10.4

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1
python-dotenv==1.0.0

# Timezone
pytz==2023.3

# CLI Framework
click==8.1.7

# Notifications
plyer==2.1.0
python-telegram-bot==20.7

# Utilities
requests==2.31.0
tabulate==0.9.0
colorama==0.4.6

# Testing
pytest==7.4.4
pytest-cov==4.1.0
